Remove defaultOptions from Tranformer class.
Instead of defaulting all transforms to ON, default them all to OFF.
This eliminates the need for defaultOptions object, because
the default options can be represented by just an empty
object.

The actual defaults used in program are defined in bin/file.js.

Other things that needed to change as a result of this:

- Explicitly enabled 'let' tranform in comments test,
  which previously relied on the tranform being enabled.

- Using simple truthiness check for options.formatter,
  as it now defaults to undefined instead of false.