Remove dead code for async file read/write
The only use of astGenerator.readFile() was to call it with
sync:true option, meaning the async-mode was never used,
which rendered the code for async file reading effectively dead.

The only use of transformer.writeFile() was to call it without
callback option, meaning it always worked in sync mode,
rendering the async code effectively dead.