Remove options param from astGenerator.read()
For some reason this.options object was passed into it, but that
function made no use of any of the options from there.  Instead
it only ever expected to see coffee option.

Decided that it should really be the task of readFile() to perform
the coffeescript compilation - this way it can simply call the
read() method without needing any options at all.

Along the way discovered that coffeescript compilation was actually
not working - coffee.compile() expected a string, but readFileSync()
return a Buffer.  Added toString conversion to fix that.