Implement CommonJS default import transform
Supporting a single var:

    var foo = require("foolib");

and a var defining multiple variables:

    var abc = 1, foo = require("foolib"), bar = require("barlib");

The code for replacing such vars is a bit tricky as ESTraverse
doesn't offer a simple way to replace a single node with multiple ones,
so I had to perform a splice() operation on parent node.

Also added tests that describe all kinds of additional cases
which we don't yet support.