Upgrade to Recast 0.11.0
This latest version now uses Esprima not Esprima-fb which has been
deprecating and was causing several parsing errors like the ones
reported in #62.

The downside of this change is that lebab will no longer handle
Facebook ES6 extensions like JSX, but that shouldn't really be
in the scope of lebab - it should be more important to properly
support full ES6 syntax.

Besides one can configure Recast to use a different parser in case
the need should arise.