Only detect CommonJS import/export in top level nodes
When CommonJS import and export is nested inside conditionals or
whatever, all kinds of trickery might be going on, and it's hard
for us to reliably detect what's going on.

ES6 modules are supposed to be defined purely statically,
so there can be no conditional importing or exporting.
Therefore it's better for us to not support such usage
at all.