Recognize Object.defineProperty() for get/set accessors
Add `kind` field to PotentialMethod.  Use it now to also
transform getters and setters.

Modify multiReplaceStatement() utility to do nothing when
the to-be-replaced node doesn't exist.  This allows us
to call .remove() on both getters, which are defined by
the same Object.defineProperty() call - the first .remove()
call will remove the Object.defineProperty(), the second
.remove() call would like to do the same, but as it's
already been removed, it does nothing.

Enable the accessors test.