Drop non-anonymous function to method conversion
This is one pretty odd feature.  I've never seen anybody defining
methods in prototype like so.  Even if one does so, it's unlikely
he wants it rewritten as this ugly delegator function.

Better to just leave the method as is - we can't easily express
this pattern with ES6 class syntax, so it's really out of scope
for Lebab to solve.

Instead convert it to test that ensures we do nothing in this case.