Do not deal with arrow-functions in class transform
I first though it was a good idea to recognize both normal
FunctionExpression and ArrowFunctionExpression, but arrow functions
can't be easily incorporated to classes when they contain
`this` references.

Instead of bothering with detecting the use of `this`
I'm simply going to ignore all arrow-functions altogether.
Pretty unlikely to encounter code where one has used
arrow-functions to construct classes in ES5-way.