Remove --module=commonjs option
Combine exportCommonjs and importCommonjs into a single transform
and enable it by default.  It can now be easily disabled if needed
by --disable=commonjs.

Because the CommonJS transform only transforms the imports and exports
that are direct children of Program node, they will not interfere
with a possible RequireJS transform we might want to write on one day
(as with RequireJS everything will be wrapped inside a function).
Therefore I think it's completely safe to enable it by default.

As a result the Lebab command line API is now really simple -
there are just --enable and --disable options for turning transforms
on/off, with all transforms being enabled by default.