Convert require("lib").foo to named import
Created additional matcher for require() with member access.
With the help of matchesAst() the matchers can now easily
share code - extracted isRequireCall() to be for matching
require("string") to be used in both matchers.

Added syntax class for ImportSpecifier and modified ImportDeclaration
to accept the specifier as input instead of hard-coding it to
ImportDefaultSpecifier.