Use _.assign() instead of Object.assign()
Our Travis build is currently configured to run with Node 0.12,
which doesn't support Object.assign().  It is supported though in
Node >= 4.0.

http://kangax.github.io/compat-table/es6/#test-Object_static_methods_Object.assign

Our package.json also defines Node version, which however is:

    "engines": {
        "node": ">= 0.10"
    }

One of these numbers needs to be changed, either we should
test on Node 0.10 as well, or say we only support node >= 0.12.
Or perhaps switch to node >= 4.0 entirely.