Remove esutils dependency
Turns out we only called one method .isExpression() from code
that was effectively dead - it set the flag hasExpression, which
was never used.  Deleted the code and the dependency.

Added some tests to ensure we properly handle all the important
expressions within string templates.

It's also good to get rid of esutils because it doesn't support ES6.