Transform named export of literal values
Convert these assignments to exported variable declarations.
This can pontentially cause conflicts when the same variable
is defined elsewhere in the file (which we don't check for).

Decided to convert these to var-s, not to let/const.  A separate
let transform can be applied if user wants to use let/const.
(The only problem right now, is that this won't happen by default,
as the commonjs transform is done after let transform.  That'll
be easy to change though.)

undefined & NaN are treated as identifers, although they are
normally used/thought-of as literals.  Anyway... these are odd
values to export anyway... but added regression tests to cover
these just in case.