Transform object destructuring to named imports
Simply converting the Properties in ObjectPattern to list
of ImportSpecifiers - a pretty simple one-to-one conversion.
Only transforming ObjectPatterns that are one level deep and
contain just identifiers as keys and values (possibly in
shorthand notation).

Changed ImportDeclaration class to take array of specifiers,
not just one.

Also did a bit of function extraction to make the varToImport()
more managable.