Add additional ESLint rules
To get most out of the linting, enabled most of ESLint rules that
I deemed useful and fit for Lebab codebase.

Some that I'd like to enable are commented out with TODO label,
as they first need some actual code to be fixed to make the linting
pass with these rules enabled.