Fix failure of arg-spread with LoDash 4.7.0
Lodash 4.7.0 fixed a bug in _.isPlainObject which the code
in arg-spread happened to rely on.  Previously _.isPlainObject
was returning true for any class instances, now it only returns
true for truly plain objects.

But the AST nodes that this code iterated over were not in fact
plain objects, but instances of some class of Recast.

Fixed by switching to _.isObjectLike() which returns true for
all "objects" except null.