Replace typeChecker with isString helper
No other functions in typeChecker are no more used, only the
isString.  So replacing the typeChecker with a single isString function.

One important difference is that the new isString()
only works on an actual AST object.  Previously one could
also pass in an undefined and have it nicely return false,
but now it crashes instead.  Only the TemplateElement relied
on this functionality - changed the code a bit to eliminate
the passing of undefined.

The isString being now a plain function, allows us to use
it more succinctly.