Initial framework for logging warnings
Created Logger class that's passed to each transform function
along with the AST.  For now none of them uses it yet.

Transformer now returns an object {code, warnings} instead of
just the string of code.  Adjusted code in tests and executable to
account for this change.  Not yet doing anything with warnings.