Use array for transforms list instead of object
We no more have by-default enabled transforms, so
there's no need for disabling them and using an
object instead of array doesn't really provide any
value.

In addition to being simpler to write,
using an arrow allows one to specify the order for
applying the transforms.  Previously the order was
pre-defined inside Lebab.

This should also pave the road for custom transforms
in the future - in addition to strings for built-in
transforms one could possibly pass in functions as well.
But that's just an idea still...

Refs #154