Rename all files from dash-separated to camelCase
So that the name of the file corresponds to the name of class or
function it contains.  For classes this also means the files are
now capitalized.

The old naming convention was a pain... one constantly needed to
switch between camelcased and dash-separated names when working
with files and functions inside them.

Consistency FTW!