Add -Test suffix to all test files
This makes it clearer which files actually contain tests.

For example we also have createTestHelpers in test/ dir which does
not contain the actual tests, but previously we still included it
among all the other test files when running Mocha.

It will also simplify searching for files as one no more needs to
look at the directory name to distinguish between the implementation
and test.