Add support for forEach transformation
This adds support for transforming simple for loops into `forEach`
loops.

This builds on the code for transforming for loops into `for-each` loops
and shares some of the `forOf` code.

The body of the loop is passed the array index only if it is used.

This will transform the following code.
```
for (let i = 0; i < fruits.length; i++) {
   let fruitItem = fruits[i];
   console.log(fruitItem);
}
// -->
fruits.forEach(fruitItem => {
   console.log(fruitItem);
});
```