Preserve get/set methods in class transform
Additional tests for ES6 methods already present in the object
literal that we're tranforming to ES6 class.

The conversion of kind=init to kind=method isn't strictly necessary,
as the Recast is able to serialase the former also to normal method,
but better safe than sorry. Always good to adhere to the AST standard.

Fixes #275